docs/legacy/original_source/subprojects/strategygen/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "strategygen"
version = "0.1.0"
description = "AI strategy generation sub-package for AISMART"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "vectorbt>=0.26",
    "pandas>=2.0",
    "pandas-ta>=0.3.14b",
    "duckdb>=0.10",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "readability-lxml>=0.8",
]

[project.optional-dependencies]
embeddings = ["sentence-transformers>=2.7"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"strategygen.ai_engine" = ["prompt_templates/*.md"]
